added matrix block extract, added moved preprocessor code into own folder in core, fixed a couple of warnings, moved properties to own subfolder
git-svn-id: https://subversion.assembla.com/svn/dsmks/trunk@600 c65b9dfa-2e9b-4d30-9b60-b5bbc3021b56

+++ src/utility.reader.ply/PlyReader.cpp
@@ -117,7 +117,7 @@ bool  PlyReader::readMesh(IMeshBuilder & builder){
     for(int i=0; i < element.instances;i++){
       mapElement(element,builder,token.stream);       
       token.next();
-      if(i%1000==0)reportProgress(token.stream.tellg());
+      if(i%1000==0)reportProgress(double(token.stream.tellg()));
       if(token.word!="\n")return false;
     }
